Add profile menu throughout the app
- "Shared with you by [...]"
- Shared with table view (in Share view)
- Comments (in Activity view)
- Activities "Shared by [...]"
