+++ iOSClient/Share/NCShareCommentsCell.swift
@@ -62,7 +62,6 @@ class NCShareCommentsCell: UITableViewCell, NCCellProtocol {
         super.awakeFromNib()
         
         buttonMenu.setImage(UIImage.init(named: "shareMenu")!.image(color: .lightGray, size: 50), for: .normal)
-        imageItem.backgroundColor = .red
         let tapGesture = UITapGestureRecognizer(target: self, action: #selector(tapAvatarImage))
         imageItem?.addGestureRecognizer(tapGesture)
     }
